+++ test/function/construction/bignumber.test.js
@@ -54,9 +54,7 @@ describe('bignumber', function() {
 
   it('should apply precision setting to bignumbers', function() {
     var math = mathjs({
-      number: {
-        precision: 32
-      }
+      decimals: 32
     });
 
     var a = math.bignumber(1).dividedBy(3);
@@ -64,9 +62,7 @@ describe('bignumber', function() {
 
     // restore default precision
     math.config({
-      number: {
-        precision: 20
-      }
+      decimals: 20
     });
   });
 });